+++ lib/paper_trail/config.rb
@@ -6,7 +6,7 @@ module PaperTrail
   # configuration can be found in `paper_trail.rb`, others in `controller.rb`.
   class Config
     include Singleton
-    attr_accessor :timestamp_field, :serializer, :version_limit
+    attr_accessor :serializer, :version_limit
     attr_writer :track_associations
 
     def initialize
@@ -15,7 +15,6 @@ module PaperTrail
       @enabled = true
 
       # Variables which affect all threads, whose access is *not* synchronized.
-      @timestamp_field = :created_at
       @serializer = PaperTrail::Serializers::YAML
     end